+++ src/utils/backend/map-settings.ts
@@ -52,8 +52,8 @@ function validateTheme(val: unknown): boolean {
 
     if ('firs' in val && !validateColor(val.firs)) return false;
     if ('uirs' in val && !validateColor(val.uirs)) return false;
-    if ('centerText' in val && !validateColor(val.uirs)) return false;
-    if ('centerBg' in val && !validateColor(val.uirs)) return false;
+    if ('centerText' in val && !validateColor(val.centerText)) return false;
+    if ('centerBg' in val && !validateColor(val.centerBg)) return false;
     if ('approach' in val && !validateColor(val.approach)) return false;
     if ('staffedAirport' in val) {
         const transparency = validateTransparency(val.staffedAirport);
@@ -108,8 +108,10 @@ const validators: Record<keyof IUserMapSettings, (val: unknown) => boolean> = {
         if ('pilots' in val && typeof val.pilots !== 'boolean') return false;
         if ('gates' in val && typeof val.gates !== 'boolean') return false;
         if ('runways' in val && typeof val.runways !== 'boolean') return false;
+        if ('pilotsInfo' in val && typeof val.pilotsInfo !== 'boolean') return false;
+        if ('atcInfo' in val && typeof val.atcInfo !== 'boolean') return false;
 
-        if (!validateRandomObjectKeys(val, ['atc', 'atcLabels', 'airports', 'pilots', 'gates', 'runways'])) return false;
+        if (!validateRandomObjectKeys(val, ['atc', 'atcLabels', 'airports', 'pilots', 'gates', 'runways', 'pilotsInfo', 'atcInfo'])) return false;
 
         return true;
     },
@@ -199,6 +201,8 @@ export interface IUserMapSettings {
         pilots?: boolean;
         gates?: boolean;
         runways?: boolean;
+        pilotsInfo?: boolean;
+        atcInfo?: boolean;
     };
     heatmapLayer: boolean;
     groundTraffic: {
@@ -261,16 +265,17 @@ export async function handleMapSettingsEvent(event: H3Event) {
             });
         }
 
+        const presets = await prisma.userPreset.findMany({
+            where: {
+                userId: user.id,
+                type: UserPresetType.MAP_SETTINGS,
+            },
+        });
+
         let settings: UserPreset | null = null;
 
         if (id) {
-            settings = await prisma.userPreset.findFirst({
-                where: {
-                    id: +id,
-                    userId: user.id,
-                    type: UserPresetType.MAP_SETTINGS,
-                },
-            }) ?? null;
+            settings = presets.find(x => x.id === +id) ?? null;
 
             if (!settings) {
                 return handleH3Error({
@@ -316,13 +321,14 @@ export async function handleMapSettingsEvent(event: H3Event) {
             }
 
             if (body.json) {
+                body.json = body.json as Record<string, any>;
+
                 for (const [key, value] of Object.entries(body.json) as [keyof IUserMapSettings, unknown][]) {
                     if (!(key in validators)) {
-                        return handleH3Error({
-                            event,
-                            statusCode: 400,
-                            statusMessage: `Invalid key given: ${ key }`,
-                        });
+                        // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
+                        delete body.json[key];
+
+                        continue;
                     }
 
                     if (!validators[key](value)) {
@@ -335,14 +341,33 @@ export async function handleMapSettingsEvent(event: H3Event) {
                 }
             }
 
-            body.json = body.json as Record<string, any>;
+            if (body.name) {
+                if (body.name.length > 20) {
+                    return handleH3Error({
+                        event,
+                        statusCode: 400,
+                        statusMessage: 'Name validation failed: max 20 symbols are allowed',
+                    });
+                }
 
-            if (body.name && body.name.length > 20) {
-                return handleH3Error({
-                    event,
-                    statusCode: 400,
-                    statusMessage: 'Name validation failed',
-                });
+                const duplicatedPreset = presets.find(x => x.name.toLowerCase().trim() === body.name?.toLowerCase().trim());
+
+                if (duplicatedPreset) {
+                    if (getQuery(event).force === '1') {
+                        await prisma.userPreset.delete({
+                            where: {
+                                id: duplicatedPreset.id,
+                            },
+                        });
+                    }
+                    else {
+                        return handleH3Error({
+                            event,
+                            statusCode: 409,
+                            statusMessage: 'A preset with this name already exists',
+                        });
+                    }
+                }
             }
 
             if (isValidate) {
@@ -358,7 +383,7 @@ export async function handleMapSettingsEvent(event: H3Event) {
                     },
                     data: {
                         name: body.name ?? settings.name,
-                        json: (body.json ?? settings.json),
+                        json: (body.json ?? settings.json ?? undefined),
                     },
                 });
             }
@@ -383,7 +408,7 @@ export async function handleMapSettingsEvent(event: H3Event) {
                         userId: user.id,
                         type: UserPresetType.MAP_SETTINGS,
                         name: body.name as string,
-                        json: body.json,
+                        json: body.json!,
                     },
                 });
             }
@@ -408,12 +433,7 @@ export async function handleMapSettingsEvent(event: H3Event) {
                 return settings;
             }
             else {
-                return prisma.userPreset.findMany({
-                    where: {
-                        userId: user.id,
-                        type: UserPresetType.MAP_SETTINGS,
-                    },
-                });
+                return presets;
             }
         }
         else {